Refactored log capture helper for outbox tests
