+++ ghost/core/test/unit/server/services/outbox/index.test.js
@@ -2,9 +2,8 @@ const assert = require('node:assert/strict');
 const sinon = require('sinon');
 const rewire = require('rewire');
 const DomainEvents = require('@tryghost/domain-events');
-const bunyan = require('bunyan');
-const {PassThrough} = require('stream');
 const logging = require('@tryghost/logging');
+const {captureLoggerOutput} = require('../../../../utils/logging-utils');
 const StartOutboxProcessingEvent = require('../../../../../core/server/services/outbox/events/start-outbox-processing-event');
 
 describe('Outbox Service', function () {
@@ -30,49 +29,6 @@ describe('Outbox Service', function () {
         await DomainEvents.allSettled();
     });
 
-    function captureLoggerOutput() {
-        const output = [];
-        const stream = new PassThrough();
-        let buffered = '';
-
-        stream.on('data', (chunk) => {
-            buffered += chunk.toString();
-            const lines = buffered.split('\n');
-            buffered = lines.pop();
-
-            for (const line of lines) {
-                if (!line.trim()) {
-                    continue;
-                }
-
-                output.push(JSON.parse(line));
-            }
-        });
-
-        const originalStreams = logging.streams;
-        logging.streams = {
-            capture: {
-                name: 'capture',
-                log: bunyan.createLogger({
-                    name: 'test-logger',
-                    streams: [{
-                        type: 'stream',
-                        stream,
-                        level: 'trace'
-                    }]
-                })
-            }
-        };
-
-        return {
-            output,
-            restore() {
-                logging.streams = originalStreams;
-                stream.destroy();
-            }
-        };
-    }
-
     describe('StartOutboxProcessingEvent subscription', function () {
         it('calls startProcessing when event is dispatched', async function () {
             service.init();
@@ -97,7 +53,7 @@ describe('Outbox Service', function () {
         });
 
         it('logs structured info if already running', async function () {
-            const capture = captureLoggerOutput();
+            const capture = captureLoggerOutput(logging);
             service.__set__('logging', logging);
 
             try {
